refactor:skinProfileService(#114)
