+++ src/main/java/com/swyp3/skin/api/v1/product/dto/response/ProductSearchResponse.java
@@ -22,7 +22,6 @@ public record ProductSearchResponse(
                 .map(product -> new ProductSummaryResponse(
                         product.getId(),
                         product.getName(),
-                        product.getBrand(),
                         product.getPrice(),
                         product.getImageUrl()
                 ))
