+++ src/main/java/com/swyp3/skin/api/v1/skintest/mapper/SkinTestResultResponseMapper.java
@@ -12,6 +12,7 @@ import com.swyp3.skin.domain.skintest.exception.SkinTestException;
 import com.swyp3.skin.recommendation.product.dto.RecommendedProduct;
 import com.swyp3.skin.recommendation.product.service.ProductRecommendationService;
 import com.swyp3.skin.recommendation.ux.IngredientMeta;
+import com.swyp3.skin.recommendation.ux.SkinProfileService;
 import com.swyp3.skin.recommendation.ux.SkinUxProfile;
 import com.swyp3.skin.recommendation.ux.SkinUxProfileResolver;
 import lombok.RequiredArgsConstructor;
@@ -27,22 +28,13 @@ import java.util.List;
 @RequiredArgsConstructor
 public class SkinTestResultResponseMapper {
 
-    public static final int PROFILE_NEED_RANKING_SIZE = 2;
-    private final SkinResultGroupScoreService skinResultGroupScoreService;
-    private final SkinUxProfileResolver skinUxProfileResolver;
-    private final ProductRecommendationService productRecommendationService;
+
+    private final SkinProfileService skinProfileService;
 
     public SkinTestResultResponse toResponse(SkinResult skinResult) {
-        List<SkinResultGroupScore> groupScores = skinResultGroupScoreService.getTop2ScoresByResultId(skinResult.getId());
-        List<IngredientGroup> ranking = groupScores.stream()
-                .map(SkinResultGroupScore::getIngredientGroup)
-                .toList();
 
-        validateRankingSize(ranking);
-        SkinUxProfile profile = resolveProfile(ranking);
-
+        SkinUxProfile profile = skinProfileService.getProfile(skinResult.getId());
         List<IngredientMeta> ingredientMetas = getIngredientMetas(profile);
-
         return SkinTestResultResponse.of(today(), profile, ingredientMetas);
     }
 
@@ -61,16 +53,5 @@ public class SkinTestResultResponseMapper {
     private static String today() {
         return LocalDate.now(ZoneId.of("Asia/Seoul"))
                 .format(DateTimeFormatter.ofPattern("yyyy.MM.dd"));
-    }
-
-    private @NonNull SkinUxProfile resolveProfile(List<IngredientGroup> ranking) {
-        return skinUxProfileResolver.resolve(ranking.get(0), ranking.get(1))
-                .orElseThrow(() -> new SkinTestException(SkinTestErrorCode.UX_PROFILE_NOT_FOUND));
-    }
-
-    private static void validateRankingSize(List<IngredientGroup> ranking) {
-        if (ranking.size() < PROFILE_NEED_RANKING_SIZE) {
-            throw new SkinTestException(SkinTestErrorCode.INVALID_RECOMMENDATION_RANKING_SIZE);
-        }
     }
 }
