+++ src/main/java/com/swyp3/skin/domain/product/service/ProductService.java
@@ -1,6 +1,7 @@
 package com.swyp3.skin.domain.product.service;
 
 import com.swyp3.skin.api.v1.product.dto.response.ProductSearchResult;
+import com.swyp3.skin.domain.common.pagination.CursorPaginationUtils;
 import com.swyp3.skin.domain.common.pagination.SliceResult;
 import com.swyp3.skin.domain.product.domain.entity.Product;
 import com.swyp3.skin.domain.product.domain.exception.ProductErrorCode;
@@ -24,15 +25,19 @@ public class ProductService {
                 .orElseThrow(() -> new ProductException(ProductErrorCode.PRODUCT_NOT_FOUND));
     }
 
-    public ProductSearchResult search(String keyword, int page, int size) {
+    public ProductSearchResult search(String keyword, int cursor, int size) {
         List<Product> productList = productRepository.search(keyword);
 
+        SliceResult<Product> productSliceResult =
+                CursorPaginationUtils.sliceWithCursor(productList, cursor, size, Product::getId);
+
+
         List<Product> sliced = productList.stream()
-                .skip((long) page * size)
+                .skip((long) cursor * size)
                 .limit(size)
                 .toList();
 
-        boolean hasNext = productList.size() > (page + 1) * size;
+        boolean hasNext = productList.size() > (cursor + 1) * size;
 
         return new ProductSearchResult(sliced, hasNext);
     }
